# server/src/swaggerDocument.json
{
  "swagger": "2.0",
  "info": {
    "version": "3.0.0",
    "title": "API DOCS",
    "description": "AUTH-APP"
  },
  "host": "localhost:8080",
  "tags": [
    {
      "name": "User",
      "description": "API for users"
    },
    {
      "name": "Register",
      "description": "Register user"
    },
    {
      "name": "Login",
      "description": "Login user"
    },
    {
      "name": "Team",
      "description": "API for teams"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/api/users": {
      "get": {
        "tags": ["User"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      }
    },
    "/api/users/:id": {
      "get": {
        "tags": ["User"],
        "parameters": [
          {
            "name": "id",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UserId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/api/register": {
      "post": {
        "tags": ["Register"],
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Register"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "tags": ["Login"],
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Token"
            }
          }
        }
      }
    },
    "/api/teams": {
      "get": {
        "tags": ["Team"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Team"
              }
            }
          }
        }
      }
    },
    "/api/teams/:teamId": {
      "get": {
        "tags": ["Team"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Team"
            }
          }
        }
      }
    },
    "/api/team": {
      "post": {
        "tags": ["Team"],
        "parameters": [
          {
            "name": "name",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/TeamName"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        }
      }
    },
    "/api/team/:teamId": {
      "put": {
        "tags": ["Team"],
        "parameters": [
          {
            "name": "name",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ModifyTeam"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        }
      },
      "delete": {
        "tags": ["Team"],
        "parameters": [
          {
            "name": "name",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ModifyTeam"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        }
      }
    },
    "/api/team/:teamId/member/:memberId": {
      "put": {
        "tags": ["Team"],
        "parameters": [
          {
            "name": "member",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/AddedMember"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        }
      },
      "delete": {
        "tags": ["Team"],
        "parameters": [
          {
            "name": "name",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/AddedMember"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "properties": {
        "id": {
          "type": "integer"
        },
        "user_name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "owned_teams": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OwnedTeam"
          }
        },
        "teams": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Team"
          }
        }
      }
    },
    "Register": {
      "required": ["user_name", "email", "password"],
      "properties": {
        "user_name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "Login": {
      "required": ["email", "password"],
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "Token": {
      "properties": {
        "access_token": {
          "type": "string"
        }
      }
    },
    "Message": {
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "Team": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "admin": {
          "$ref": "#/definitions/Member"
        },
        "members": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Member"
          }
        }
      }
    },
    "OwnedTeam": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "members": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Member"
          }
        }
      }
    },
    "Member": {
      "properties": {
        "id": {
          "type": "integer"
        },
        "user_name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
    "TeamName": {
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "ModifyTeam": {
      "properties": {
        "name": {
          "type": "string"
        },
        "id": {
          "type": "string"
        }
      }
    },
    "AddedMember": {
      "properties": {
        "teamId": {
          "type": "string"
        },
        "memberId": {
          "type": "string"
        }
      }
    },
    "UserId": {
      "properties": {
        "id": {
          "type": "string"
        }
      }
    }
  }
}